integrando o resto dos services do profissional e do user em seus respctivos Controllers, criando interfaces e services das classes de Plantao e Setor e integrando em seus Controllers e outras pequenas alterações

+++ MeuPlantao.Domain/Entities/TrocaHistoricoModel.cs
@@ -5,12 +5,12 @@ namespace MeuPlantao.Domain.Entities;
 
 public class TrocaHistoricoModel
 {
-    public int Id { get; set; }
-    public int TrocaPlantaoId { get; set; }
+    public long Id { get; set; }
+    public long TrocaPlantaoId { get; set; }
     public TrocaPlantaoModel TrocaPlantao { get; set; } = null!;
     public EventoHistoricoEnum Evento { get; set; }
     public DateTime Data { get; set; } = DateTime.UtcNow;
-    public int UsuarioId { get; set; }
+    public long UsuarioId { get; set; }
     public UserModel Usuario { get; set; } = null!;
     [MaxLength(100)]
     public string Observacao { get; set; } = string.Empty;
